move MajorDefensives components to core, parameterize
